+++ src/api/index.ts
@@ -1,4 +1,8 @@
 export * from './relay';
-export * from './route';
 export * from './testTimeout';
 export * from './getVersion';
+export * from './wormhole';
+export * from './xcm';
+export * from './homa';
+export * from './euphrates';
+export * from './health';
